requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
cairocffi==1.7.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
Django==5.2.3
django-countries==7.6.1
django-environ==0.12.0
django-phonenumber-field==8.1.0
fonttools==4.58.5
html5lib==1.1
idna==3.10
lxml==6.0.0
oscrypto==1.3.0
phonenumbers==9.0.9
pillow==11.2.1
psycopg2-binary==2.9.10
pycairo==1.28.0
pycparser==2.22
pydyf==0.11.0
##PyGObject==3.52.3
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pypdf==5.8.0
pyphen==0.17.2
python-bidi==0.6.6
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
django-widget-tweaks==1.5.0
